+++ src/ui/gui/videoviewfactory.h
@@ -20,15 +20,15 @@ public:
   // conferenceview is needed for connecting reattach signal, because I couldn't get the
   // the interface signal connected for some reason.
   // returns viewID which is also videoID
-  uint32_t createWidget(uint32_t sessionID, QWidget* parent, ConferenceView* conf);
+  size_t createWidget(uint32_t sessionID, QWidget* parent, ConferenceView* conf);
 
   // set self view which is part of the
   // returns viewID which is also videoID
-  uint32_t setSelfview(VideoInterface* video, QWidget* view);
+  size_t setSelfview(VideoInterface* video, QWidget* view);
 
   // 0 in sessionID is for selfview
   // id is the index of that view or video
-  QWidget* getView(uint32_t sessionID, uint32_t viewID);
+  QWidget* getView(uint32_t sessionID, size_t viewID);
   VideoInterface* getVideo(uint32_t sessionID, uint32_t videoID);
 
   // Does not clear selfview
